app/utils/jwt_handler.py
from jose import jwt
from datetime import datetime, timedelta

# Can be added to the env but for assigment purposes written here itself
SECRET = "SUPERSECRET"
ALGORITHM = "HS256"

def create_token(data: dict):
    payload = data.copy()
    payload["exp"] = datetime.utcnow() + timedelta(hours=2)
    return jwt.encode(payload, SECRET, algorithm=ALGORITHM)

def decode_token(token: str):
    return jwt.decode(token, SECRET, algorithms=[ALGORITHM])