requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-multipart